A-MoreOOP: Added Storage Class, pending Parser
