+++ src/main/java/Duke.java
@@ -1,75 +1,9 @@
-
 import java.util.Scanner;
-import java.util.Arrays;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.io.File;
-import java.io.FileNotFoundException;
 import java.text.ParseException;
 import subclass.*;
 
 public class Duke {
-
-    public static void writeToFile(String filePath, String textToAdd) throws IOException {
-        FileWriter fw = new FileWriter(filePath);
-        fw.write(textToAdd);
-        fw.close();
-    }
-
-    public static void loadFile(String filePath) throws FileNotFoundException, ParseException {
-        File f = new File(filePath);
-        Scanner s = new Scanner(f);
-        int num = 1;
-        while (s.hasNext()) {
-            String line_input = s.nextLine();
-            //add todo, if any
-            if (line_input.contains("T|")) {
-                String input_items = line_input.substring(line_input.lastIndexOf("|") + 1);
-                String task_status_num = line_input.substring(line_input.indexOf("|") + 1);
-                Task.add_task(new Todo(input_items));
-                if (task_status_num.contains("1")){
-                    Task.markDone(num);
-                }
-            }
-            //add deadline, if any
-            if (line_input.contains("D|")) {
-                String input_by = line_input.substring(line_input.lastIndexOf("|") + 1);
-                String input_items = line_input.substring(line_input.indexOf("|") + 3, line_input.lastIndexOf("|"));
-                String task_status_num = line_input.substring(line_input.indexOf("|") + 1);
-                Task.add_task(new Deadline(input_items, input_by));
-                if (task_status_num.contains("1")){
-                    Task.markDone(num);
-                }
-            }
-            //add event, if any
-            if (line_input.contains("E|")) {
-                String input_by = line_input.substring(line_input.lastIndexOf("|") + 1);
-                String input_items = line_input.substring(line_input.indexOf("|") + 3, line_input.lastIndexOf("|"));
-                String task_status_num = line_input.substring(line_input.indexOf("|") + 1);
-                Task.add_task(new Event(input_items, input_by));
-                if (task_status_num.contains("1")){
-                    Task.markDone(num);
-                }
-            }
-            num++;
-        }
-    }
-
-    public static String toTxt (String outputs) {
-        String newOutput = outputs.replace("[", "");
-        newOutput = newOutput.replace("]", ",");
-        newOutput = newOutput.replace("(", ",");
-        newOutput = newOutput.replace(")", "");
-        newOutput = newOutput.replace("\u2713", "1");
-        newOutput = newOutput.replace("\u2718", "0");
-        newOutput = newOutput.replace("by:", "");
-        newOutput = newOutput.replace("at:", "");
-        newOutput = newOutput.replace(", ", ",");
-        newOutput = newOutput.replace("  ,", ",");
-        newOutput = newOutput.replace(",", "|");
-        newOutput = newOutput.replace(" |", "|");
-        return newOutput;
-    }
 
     public static void main(String[] args) throws DukeException, todoException, IOException, ParseException {
 
@@ -77,7 +11,7 @@ public class Duke {
         Scanner in = new Scanner(System.in);
         String file_path = "C:\\Users\\marcus.ng\\Desktop\\m\\m\\NUS\\TIC2002 Introduction to Software Engineering\\duke\\src\\main\\java\\taskList.txt";
         try {
-            loadFile(file_path);
+            Storage.loadFile(file_path);
         } catch (ParseException e) {
             Ui.displayError();
         }
@@ -101,7 +35,6 @@ public class Duke {
                         Ui.showLine();
                         Ui.displayGotIt();
                         System.out.println("\t\t" + newTask.add_task(new Todo(input_items)));
-                        //word_count++;
                         System.out.println("\t\tNow you have " + newTask.word_count + " tasks in list.");
                         Ui.showLine();
 
@@ -193,7 +126,7 @@ public class Duke {
 
             }
         //save Task.txt
-        writeToFile(file_path, toTxt(Task.getOutput()));
+        Storage.writeToFile(file_path, Storage.toTxt(Task.getOutput()));
 
         //exit
         Ui.showLine();
