+++ src/main/java/taskList.txt
@@ -1,6 +1,7 @@
 T|1|read books 
-D|0|do accounts|19-Jul-19 06:30 pm
+D|1|do accounts|19-Jul-19 06:30 PM
 E|0|meeting|Suntec 8pm
 T|1|swimming
 E|0|boxing match|Gym
-D|0|do math|23-Sep-19 08:00 pm
+D|1|do math|23-Sep-19 08:00 PM
+T|1|basketball practice
